+++ CHANGELOG.md
@@ -4,6 +4,8 @@
 
 ## [Unreleased] - ReleaseDate
 
+## [1.1.1] - 2025-08-09
+
 ### Added
 
 - Availability sensor, which tells Home Assistant when the device is connected/disconnected.
@@ -37,7 +39,8 @@ Initial version
 
 
 <!-- next-url -->
-[Unreleased]: https://github.com/guillaumecl/mqtt-system-monitor/compare/v1.1.0...HEAD
+[Unreleased]: https://github.com/guillaumecl/mqtt-system-monitor/compare/v1.1.1...HEAD
+[1.1.1]: https://github.com/guillaumecl/mqtt-system-monitor/compare/v1.1.0...v1.1.1
 [1.1.0]: https://github.com/guillaumecl/mqtt-system-monitor/compare/v1.0.5...v1.1.0
 [1.0.5]: https://github.com/guillaumecl/mqtt-system-monitor/compare/v1.0.3...v1.0.5
 [1.0.3]: https://github.com/guillaumecl/mqtt-system-monitor/releases/tag/v1.0.3